Bunch of changes

+++ src/main/java/org/orecruncher/sndctrl/audio/BackgroundSoundInstance.java
@@ -47,17 +47,17 @@ public class BackgroundSoundInstance extends FadableSoundInstance {
     }
 
     @Override
-    public float getX() {
+    public double getX() {
         return 0;
     }
 
     @Override
-    public float getY() {
+    public double getY() {
         return 0;
     }
 
     @Override
-    public float getZ() {
+    public double getZ() {
         return 0;
     }
